Added Gem Functionality
Keys, rocks, and gems now work as intended and have prefabs. The Gem script takes a door that can be opened and closed by stepping on and off the switch or dropping rocks onto it. Rocks and Keys can be set down with the "E" key, the player must walk off of the object and touch it again to pick it back up after dropping it.